+++ src/main/java/com/litesoftwares/coingecko/domain/Coins/CoinMarkets.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.litesoftwares.coingecko.domain.Coins.CoinData.Roi;
 import com.litesoftwares.coingecko.domain.Coins.CoinData.SparklineIn7d;
+import java.math.BigDecimal;
 import lombok.*;
 
 @Data
@@ -18,35 +19,35 @@ public class CoinMarkets {
     @JsonProperty("image")
     private String image;
     @JsonProperty("current_price")
-    private double currentPrice;
+    private BigDecimal currentPrice;
     @JsonProperty("market_cap")
-    private double marketCap;
+    private BigDecimal marketCap;
     @JsonProperty("market_cap_rank")
     private long marketCapRank;
     @JsonProperty("fully_diluted_valuation")
-    private double fullyDilutedValuation;
+    private BigDecimal fullyDilutedValuation;
     @JsonProperty("total_volume")
-    private double totalVolume;
+    private BigDecimal totalVolume;
     @JsonProperty("high_24h")
-    private double high24h;
+    private BigDecimal high24h;
     @JsonProperty("low_24h")
-    private double low24h;
+    private BigDecimal low24h;
     @JsonProperty("price_change_24h")
-    private double priceChange24h;
+    private BigDecimal priceChange24h;
     @JsonProperty("price_change_percentage_24h")
-    private double priceChangePercentage24h;
+    private BigDecimal priceChangePercentage24h;
     @JsonProperty("market_cap_change_24h")
-    private double marketCapChange24h;
+    private BigDecimal marketCapChange24h;
     @JsonProperty("market_cap_change_percentage_24h")
-    private double marketCapChangePercentage24h;
+    private BigDecimal marketCapChangePercentage24h;
     @JsonProperty("circulating_supply")
-    private double circulatingSupply;
+    private BigDecimal circulatingSupply;
     @JsonProperty("total_supply")
-    private double totalSupply;
+    private BigDecimal totalSupply;
     @JsonProperty("ath")
-    private double ath;
+    private BigDecimal ath;
     @JsonProperty("ath_change_percentage")
-    private double athChangePercentage;
+    private BigDecimal athChangePercentage;
     @JsonProperty("ath_date")
     private String athDate;
     @JsonProperty("roi")
@@ -56,7 +57,7 @@ public class CoinMarkets {
     @JsonProperty("sparkline_in_7d")
     private SparklineIn7d sparklineIn7d;
     @JsonProperty("price_change_percentage_1h_in_currency")
-    private double priceChangePercentage1hInCurrency;
+    private BigDecimal priceChangePercentage1hInCurrency;
 
 }
